Added capability to cross compile
The original rebar.config.script uses erland:system_info(system_architecture)
to obtain the machine triplet e.g x86_64-redhat-linux-gnu. This does not work
in a cross environment where rebar runs inside a native Erlang runtime. The
rebar script now tries to detect if it runs in a cross environment by checking
the envrionment variable CXX. If it's set that typically indicates that a
cross compiler is used (but even if not it won't hurt). It then determines the
machine triplet from the cross compiler as well as the sysroot.

Signed-off-by: Rudolf J Streif <rudolf.streif@gmail.com>